Mobile Test Case -->   End of Appium Mobile Section

+++ src/main/java/utilities/Listeners.java
@@ -1,5 +1,7 @@
 package utilities;
 
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.screenrecording.CanRecordScreen;
 import io.qameta.allure.Allure;
 import io.qameta.allure.Attachment;
 import org.openqa.selenium.OutputType;
@@ -10,64 +12,83 @@ import org.testng.ITestListener;
 import org.testng.ITestResult;
 
 import java.io.File;
+import java.util.Date;
 
 
 public class Listeners extends CommonOps implements ITestListener {
 
+
     public void onStart (ITestContext execution) {
-        System.out.println("---------------------- Starting Execution ------------------");
+        System.out.println(execution.getStartDate() +" :------------------- Starting Execution ------------------");
     }
 
     public void onFinish (ITestContext execution) {
-        System.out.println("---------------------- Ending Execution ------------------");
+        System.out.println(" :---------------------- Ending Execution ------------------");
+        System.out.println("finished : " + execution.getEndDate());
     }
 
     public void onTestSkipped (ITestResult test) {
-        System.out.println("---------------------- Skipping Test "  + test.getName() + "  ------------------");
+        System.out.println(" :---------------------- Skipping Test " + test.getName() + " -----------------" );
+
     }
 
     public void onTestStart (ITestResult test) {
-        System.out.println("---------------------- Starting Test: " + test.getName() + "  ------------------");
+        System.out.println(" ---------------------- Starting Test: " + test.getName() + "  ------------------");
     }
 
     public void onTestSuccess (ITestResult test) {
-        System.out.println("---------------------- Test: " + test.getName() + " Passed ------------------");
+        System.out.println(" ----------  SUCCESS! ----------- Test: " + test.getName() + " Passed ------------------");
 
-        // Stop recording
-        try {
-            MonteScreenRecorder.stopRecord();
-        } catch (Exception e){
-            e.printStackTrace();
+        if (isWeb()) {
+            try {
+                MonteScreenRecorder.stopRecord();
+//                   ((CanRecordScreen) mobileDriver).stopRecordingScreen();
+            } catch (Exception e){
+                e.printStackTrace();
+            }
         }
 
-        // Delete recorded file
 
-        File file = new File("./test-recordings/" + test.getName() + ".avi");
-        if (file.delete())
-            System.out.println("File was deleted successfully");
-        else
-            System.out.println("Failed to delete the file");
+        if (isWeb()) {
+            // Delete recorded file
+            File file = new File("./test-recordings/" + test.getName() + ".avi");
+            if (file.delete())
+                System.out.println("File was deleted successfully");
+            else
+                System.out.println("Failed to delete the file");
+        }
     }
 
     public void onTestFailure (ITestResult test) {
+
         System.out.println("---------------------- Test "  + test.getName() + " Failed ------------------");
         try {
-            MonteScreenRecorder.stopRecord();
+            if (isWeb())
+                MonteScreenRecorder.stopRecord();
         } catch (Exception e){
+            System.out.println("Error recording video of screen :" + e);
             e.printStackTrace();
         }
-        saveScreenshot(driver);
+        System.out.println(new Date(test.getStartMillis()  ) + test.getName() + "SCREENSHOT TAKEN : ");
+
+        saveScreenshot();
     }
 
     public void onTestFailedButWithinSuccessPercentage(ITestResult arg0) {
     }
+
     @Attachment(value = "Page Screen Shot" ,type = "image/png")
-    public byte[] saveScreenshot (WebDriver driver) {
+    public byte[] saveScreenshot () {
 //        Allure.addAttachment(test.getName(), new ByteArrayInputStream(((TakesScreenshot)driver).getScreenshotAs(OutputType.BYTES)));
 //        TakesScreenshot ts = (TakesScreenshot)driver;
 //        File file = ts.getScreenshotAs(OutputType.FILE);
 //        FileUtils.copyFile(file, new File("./ScreenShot_Folder/TestFailure.png"));
-        return ((TakesScreenshot)driver).getScreenshotAs(OutputType.BYTES);
+
+        if (isMobile())
+            return ((TakesScreenshot)mobileDriver).getScreenshotAs(OutputType.BYTES);
+        else
+            return ((TakesScreenshot)driver).getScreenshotAs(OutputType.BYTES);
+
     }
 
 }
